+++ Tree.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2013
 VisualStudioVersion = 12.0.31101.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tree", "TestTree\Tree.csproj", "{66202F7D-96E6-4AD1-A6A4-CE2292294140}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnitTestTree", "UnitTestTree\UnitTestTree.csproj", "{A0C66A52-EE46-444B-BA1A-D1FF3FE76BED}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tree", "Tree\Tree.csproj", "{66202F7D-96E6-4AD1-A6A4-CE2292294140}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,14 +13,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{66202F7D-96E6-4AD1-A6A4-CE2292294140}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{66202F7D-96E6-4AD1-A6A4-CE2292294140}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{66202F7D-96E6-4AD1-A6A4-CE2292294140}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{66202F7D-96E6-4AD1-A6A4-CE2292294140}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A0C66A52-EE46-444B-BA1A-D1FF3FE76BED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A0C66A52-EE46-444B-BA1A-D1FF3FE76BED}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A0C66A52-EE46-444B-BA1A-D1FF3FE76BED}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A0C66A52-EE46-444B-BA1A-D1FF3FE76BED}.Release|Any CPU.Build.0 = Release|Any CPU
+		{66202F7D-96E6-4AD1-A6A4-CE2292294140}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{66202F7D-96E6-4AD1-A6A4-CE2292294140}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{66202F7D-96E6-4AD1-A6A4-CE2292294140}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{66202F7D-96E6-4AD1-A6A4-CE2292294140}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
