+++ Tree/Point.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace TestTree
+namespace Tree
 {
     public class Point
     {
